IGDD-1954 / IGDD-1953 - Access Control and User API (#130)
* Add .run folder (create by IntellJ depending on configuration) to .gitignore.

* Add CRUD calls for AccessControl to ApiController.  Update validation messages on AccessControl model.

* Add user validation so that checks can be done in Access Control API that a valid userId is being passed.

* Add CRUD calls for User to ApiController.

* Update User model with more validations and updated validation messages.  Check in users.json default configuration with space/syntax to match how it gets updated via API.

* Fix validation message in User model.  Add Postman tests for Users api.  Start of Postman tests for Access Controls.

* Update spacing/syntax on default access-control.json file to match format set by processing via API.

* Finish Access Control tests in Postman.

* Create BaseApiController as start of splitting monster ApiController.  Create UserApiController.

* Create PipelineApiController

* Split the rest of the API endpoints into their own files.  Removed ApiController.  Tweaks to postman for some access control tests to work (needed to look for a specific validation error inside the message instead of assuming the whole message would be equal).

* Tweak Access Control postman tests so that the temporary Organization and User needed are always added as active.  Was randomly picking a boolean value, resulting in false errors when the User and/or Organization were added active = false.

* Removed ability to delete User via API after internal discussion.  Update Postman tests to account for the idea that we won't be able to add a temporary user, both to test the User endpoints but also the Access Control test.
