+++ .gitignore
@@ -18,6 +18,7 @@ target/
 *.iws
 *.iml
 *.ipr
+/.run/
 
 ### NetBeans ###
 /nbproject/private/
