+++ src/main/java/gov/cdc/izgateway/xform/model/User.java
@@ -1,5 +1,6 @@
 package gov.cdc.izgateway.xform.model;
 
+import gov.cdc.izgateway.xform.validation.ValidOrganization;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
@@ -10,12 +11,17 @@ import java.util.UUID;
 @Getter
 @Setter
 public class User implements BaseModel {
-    @NotBlank(message = "User name is required")
+    @NotBlank(message = "User - Name is required")
     private String userName;
+
+    @NotNull(message = "User - ID is required")
     private UUID id;
 
-    // The organization associated with the user - null means the user has no organization affiliation
+    @NotNull(message = "User - Organization is required")
+    @ValidOrganization(message = "Organization ID must reference an existing and active organization")
     private UUID organizationId;
+
+    @NotNull(message = "User - Active status is required")
     private Boolean active;
 
     public User() {
