+++ testing/configuration/users.json
@@ -1,38 +1,31 @@
-[
-  {
-    "id": "4f5b1f8e-6c7d-4a9b-8e3f-2d7c6b5a9e4d",
-    "userName": "kboone.testing.izgateway.org",
-    "organizationId": "f56a10cc-bf37-4257-bba7-44dcf376dc3e",
-    "active": true
-  },
-  {
-    "id": "8c942339-e10d-418f-8898-d8e73388198c",
-    "userName": "pcahill.testing.izgateway.org",
-    "organizationId": "0d15449b-fb08-4013-8985-20c148b353fe",
-    "active": true
-  },
-  {
-    "id": "f1d0616c-ec7c-4070-b43a-9e7295d45d21",
-    "userName": "amoody.testing.izgateway.org",
-    "organizationId": "6d531709-6911-463f-9ee8-5fd704f22b24",
-    "active": true
-  },
-  {
-    "id": "61d7e381-2807-4d0f-8e6a-8a0df182d70d",
-    "userName": "bharris.testing.izgateway.org",
-    "organizationId": "a9bc153a-d400-42e7-8e74-2e7d8b2e1481",
-    "active": true
-  },
-  {
-    "id": "7ae8e29b-aca6-4795-8de6-1183236e56b6",
-    "userName": "ppatel.testing.izgateway.org",
-    "organizationId": "8c905096-c79c-4351-86a7-1d6b7da5ec44",
-    "active": true
-  },
-  {
-    "id": "8556e4a3-743a-4e71-9971-6d8e2d679ed6",
-    "userName": "preprod-cc.phiz-project.org",
-    "organizationId": "11d32c29-b91a-455f-931e-591cb8ee4f51",
-    "active": true
-  }
-]
+[ {
+  "userName" : "kboone.testing.izgateway.org",
+  "id" : "4f5b1f8e-6c7d-4a9b-8e3f-2d7c6b5a9e4d",
+  "organizationId" : "f56a10cc-bf37-4257-bba7-44dcf376dc3e",
+  "active" : true
+}, {
+  "userName" : "pcahill.testing.izgateway.org",
+  "id" : "8c942339-e10d-418f-8898-d8e73388198c",
+  "organizationId" : "0d15449b-fb08-4013-8985-20c148b353fe",
+  "active" : true
+}, {
+  "userName" : "amoody.testing.izgateway.org",
+  "id" : "f1d0616c-ec7c-4070-b43a-9e7295d45d21",
+  "organizationId" : "6d531709-6911-463f-9ee8-5fd704f22b24",
+  "active" : true
+}, {
+  "userName" : "bharris.testing.izgateway.org",
+  "id" : "61d7e381-2807-4d0f-8e6a-8a0df182d70d",
+  "organizationId" : "a9bc153a-d400-42e7-8e74-2e7d8b2e1481",
+  "active" : true
+}, {
+  "userName" : "ppatel.testing.izgateway.org",
+  "id" : "7ae8e29b-aca6-4795-8de6-1183236e56b6",
+  "organizationId" : "8c905096-c79c-4351-86a7-1d6b7da5ec44",
+  "active" : true
+}, {
+  "userName" : "preprod-cc.phiz-project.org",
+  "id" : "8556e4a3-743a-4e71-9971-6d8e2d679ed6",
+  "organizationId" : "11d32c29-b91a-455f-931e-591cb8ee4f51",
+  "active" : true
+} ]
